Updating PyProject.toml
