+++ eng/templates/jobs/ci-unit-tests.yml
@@ -32,9 +32,10 @@ jobs:
             python -m pip install --pre -U -e .[test-http-v2]
           fi
 
-          python setup.py build
-          python setup.py webhost --branch-name=dev
-          python setup.py extension
+          cd tests
+          python -m invoke -c test_setup build-protos
+          python -m invoke -c test_setup webhost --branch-name=dev
+          python -m invoke -c test_setup extension
         displayName: "Install dependencies"
       - bash: |
              python -m pytest -q -n auto --dist loadfile --reruns 4 --instafail --cov=./azure_functions_worker --cov-report xml --cov-branch tests/unittests
