+++ eng/templates/official/jobs/ci-docker-consumption-tests.yml
@@ -56,7 +56,8 @@ jobs:
           if [[ $(PYTHON_VERSION) != "3.8" ]]; then
               python -m pip install --pre -U -e .[test-deferred-bindings]
           fi
-          python setup.py build
+          cd tests
+          python -m invoke -c test_setup build-protos
         displayName: 'Install dependencies'
       - bash: |
              python -m pytest --reruns 4 -vv --instafail tests/endtoend tests/extension_tests/deferred_bindings_tests tests/extension_tests/http_v2_tests
