+++ eng/templates/official/jobs/ci-e2e-tests.yml
@@ -74,9 +74,10 @@ jobs:
               python -m pip install --pre -U -e .[test-deferred-bindings]
           fi
 
-          python setup.py build
-          python setup.py webhost --branch-name=dev
-          python setup.py extension
+          cd tests
+          python -m invoke -c test_setup build-protos
+          python -m invoke -c test_setup webhost --branch-name=dev
+          python -m invoke -c test_setup extension
         displayName: 'Install dependencies and the worker'
         condition: and(eq(variables['USETESTPYTHONSDK'], false), eq(variables['USETESTPYTHONEXTENSIONS'], false))
       - task: DownloadPipelineArtifact@2
@@ -101,9 +102,10 @@ jobs:
               python -m pip install --pre -U -e .[test-deferred-bindings]
           fi
 
-          python setup.py build
-          python setup.py webhost --branch-name=dev
-          python setup.py extension
+          cd tests
+          python -m invoke -c test_setup build-protos
+          python -m invoke -c test_setup webhost --branch-name=dev
+          python -m invoke -c test_setup extension
         displayName: 'Install test python sdk, dependencies and the worker'
         condition: eq(variables['USETESTPYTHONSDK'], true)
       - task: DownloadPipelineArtifact@2
@@ -130,9 +132,10 @@ jobs:
           
           python -m pip install -U -e .[dev]
 
-          python setup.py build
-          python setup.py webhost --branch-name=dev
-          python setup.py extension
+          cd tests
+          python -m invoke -c test_setup build-protos
+          python -m invoke -c test_setup webhost --branch-name=dev
+          python -m invoke -c test_setup extension
         displayName: 'Install test python extension, dependencies and the worker'
         condition: eq(variables['USETESTPYTHONEXTENSIONS'], true)
       - bash: |
